Evolve datatip API to allow providers to update datatip values.
Summary: Previously, a Datatip provider would pass a `ReactElement`. In order for the provider to update the content of a datatip, it needs a way to update the element after the fact. At the same time, matthewwithanm pointed out that the Datatip service might want to inject props into the provided component. You see the resulting solution here: Providers provide an observable and a component class, and the datatip service takes care of wiring them up.

Reviewed By: johnislarry

Differential Revision: D3129245

fb-gh-sync-id: a0df283aa00aac4f95f4ab77d6412d55b06ba29f
fbshipit-source-id: a0df283aa00aac4f95f4ab77d6412d55b06ba29f