Unsubscribe datatip provider on package deactivation
Summary: In D3130517, I inverted the consumer/provider relationship of the datatip service. As a result of that, toggling the datatip package on & off works correctly. However, each provider also needs to unregister itself when its package deactivates. This diff is part of a stack that updates all datatip providers accordingly.

Reviewed By: ssorallen

Differential Revision: D3135282

fb-gh-sync-id: ad6aa2ae05c45779ff301305287a7b65ddbdeba3
fbshipit-source-id: ad6aa2ae05c45779ff301305287a7b65ddbdeba3