Unsubscribe datatip provider on package deactivation
Summary: In D3130517, I inverted the consumer/provider relationship of the datatip service. As a result of that, toggling the datatip package on & off works correctly. However, each provider also needs to unregister itself when its package deactivates. This diff is part of a stack that updates all datatip providers accordingly.

Reviewed By: ssorallen

Differential Revision: D3135291

fb-gh-sync-id: fe3a2326d372f833076747893926d8f348ea8a8e
fbshipit-source-id: fe3a2326d372f833076747893926d8f348ea8a8e