Unsubscribe datatip provider on package deactivation
Summary: In D3130517, I inverted the consumer/provider relationship of the datatip service. As a result of that, toggling the datatip package on & off works correctly. However, each provider also needs to unregister itself when its package deactivates. This diff is part of a stack that updates all datatip providers accordingly.

Reviewed By: ssorallen

Differential Revision: D3135303

fb-gh-sync-id: 724a5018818dd6aba3d930d04da692c0d5ab2c7b
fbshipit-source-id: 724a5018818dd6aba3d930d04da692c0d5ab2c7b