Fix debugger instance dispose race condition
Summary:In startDebugging() we call stopDebugging() first to dispose previous session. This has a race condition that previous session may be disposed very late when the new session is alive. Then the _handleSessionEnd() will incorrectly kill the new session.
To fix it, we should only kill debugger if the ending session is the same as current live session.

Reviewed By: johnislarry

Differential Revision: D3164596

fb-gh-sync-id: 754e712e50945c842a6644c0115e62f0053fe9e8
fbshipit-source-id: 754e712e50945c842a6644c0115e62f0053fe9e8