Clang: display type tooltips for local variable declarations
Summary:Local variable declarations are often typed with `auto`, which means that type tooltips actually become useful - enable them for this case.

This does mean we'll get some useless self-linking hyperclick results for local vars as well. This doesn't seem like a big enough issue to be worth going out of the way to fix, though.

Reviewed By: yinghuitan

Differential Revision: D3165204

fb-gh-sync-id: 7a50bce6c18dfdc3d6aac2f2a2c7000b325450da
fbshipit-source-id: 7a50bce6c18dfdc3d6aac2f2a2c7000b325450da