Added file tree filtering and navigation
Summary:This change set adds new functionality that replaces the old prefix functionality.
When you click on the file tree and then type, matching terms are highlighted in blue. This only filters for currently visible/ known file-tree nodes. To navigate these matches you can use tab and shift-tab. The backspace key now removes a character from the search.

Additionally, the first highlighted result is selected when you finish typing. When a file is opened the search filter is cleared.

Reviewed By: shushz

Differential Revision: D3155230

fb-gh-sync-id: 2174c306f03ee96320da7fb00ed11972cfa9c11f
fbshipit-source-id: 2174c306f03ee96320da7fb00ed11972cfa9c11f