Fix Working Sets editing
Summary:In recent refactoring, among other things, the Working Set editing was optimized to take advantage of the new hierarchical nature of the `FileTreeStore`. Unfortunately this optimization proven to be very hard to implement correctly for the case when editing an existing working set with potentially yet unfetched URIs.
This diff removes this optimization and makes the edited working set, as previously, a global instance for the file-tree. The `checkedStatus` property of the FileTreeNode is no longer settable, but rather always derived from this global instance.

Fixes #454

Reviewed By: ssorallen

Differential Revision: D3174889

fb-gh-sync-id: 0191d09fc1cb77d85e2ade372652ac752d273ef0
fbshipit-source-id: 0191d09fc1cb77d85e2ade372652ac752d273ef0