Clang: Detect changes in BUCK and TARGETS files (1/2)
Summary:This augments `ClangFlagsManager` to install file watchers on build files (BUCK, TARGETS files) and `compile_commands.json` files.

Clients asking for flags will additionally receive an `onChange` Observable which they can use to listen for changes. When a change in a build file is detected, the change event is forwarded through the Observable.

Reviewed By: matthewwithanm

Differential Revision: D3092203

fb-gh-sync-id: a1912e80e6df1a765e0379a260fef84fd05bb619
fbshipit-source-id: a1912e80e6df1a765e0379a260fef84fd05bb619