first diff for wootr library
Summary:Introduce main types and logic for inserting "runs" or ranges of WChars.  The idea is that if the degree and local clock id (h) is just one more than the previous character, and the visibility property is the same, then we can merge the two characters into 1 run.  As long as people type/edit mostly sequentially we should merge the majority of the time.

Each WCharRun is an ID, boolean for visibility, degree, and length.  The ID is a site and h (local clock).

Reviewed By: nmote

Differential Revision: D3043193

fb-gh-sync-id: a3db4ec87846bfff06556c44e7a93bea14b17283
fbshipit-source-id: a3db4ec87846bfff06556c44e7a93bea14b17283