Fix uncaught promise error when starting debugging
Summary:If we restart debugging then this rejecting promise is actually handled in _waitForChromeConnection,
or near there, but if we are debugging for the first time then the promise is constructed in the constructor
of the store, so we have to catch a rejection there as well.  Doing this gets rid of the uncaught promise
error in the console.

Reviewed By: yinghuitan

Differential Revision: D3176051

fb-gh-sync-id: ef02606bef73239564ac36cb792c9a856eeea7c8
fbshipit-source-id: ef02606bef73239564ac36cb792c9a856eeea7c8