Publish View UX: Replace nuclide-console with inline publish updates editor
Summary:The current nuclide-console usage is a terrible UX experience where the Diff View is squeezed to half,
while the console is taking the other half of the screen.
That leaves the readonly publish message text editor area unused.
This diff reuses this area to show the real-time updates of `arc diff` inline.

Reviewed By: matthewwithanm

Differential Revision: D3187526

fb-gh-sync-id: c06cac429f5968a807ae332ff933d5d1275d0096
fbshipit-source-id: c06cac429f5968a807ae332ff933d5d1275d0096