Add Nuclide require profiler
Summary: From `pre-main.js` pass `track-nuclide-ready.js#init` whether you want to print to the console a require/load-time profile for all of nuclide (`byLoadOrder`) or by feature (`byRoot`). By default, in development, only `byRoot` is shown ? when packaged, these files don't even run. The durations are somewhat misleading because they represent "how long did it take to require this module?", and not 'how long did it take to load this module plus all of the modules it loaded during the profile". Basically, this means that deferred requires aren't reflected in the parent modules load time.

Reviewed By: nmote, jamesgpearce

Differential Revision: D3190454

fb-gh-sync-id: 1c2d04615b6c1aca59320593b36954e44ea61224
fbshipit-source-id: 1c2d04615b6c1aca59320593b36954e44ea61224