Use React.{Component,Element} instead of ReactComponent/ReactElement types
Summary:These changes are compatible with flow 0.21.0 and flow 0.23.0- in preparation for an update to flow 0.23.0.

I'm not sure if it's because we go through `react-for-atom`, but `import type {React} from 'react-for-atom';` doesn't work. But since we load `react-for-atom` at startup (in even the most slim setup), there's no perf harm to importing `react-for-atom` just for type info.

Reviewed By: nmote

Differential Revision: D3186827

fb-gh-sync-id: 06f1b5b6fb9968b98184b48c0f9f544977decbdb
fbshipit-source-id: 06f1b5b6fb9968b98184b48c0f9f544977decbdb