Introduce development only atom-babel-compiler-patcher
Summary:`atom-babel-compiler-patcher` monkey-patches Atom's babel compiler so that we can apply custom transforms on Nuclide code only. Currently, the only differences in transpilation between this and Atom, is that we're not transpiling template strings and we're not including source maps.

This only runs during development - at packing time, `pre-main.js` doesn't run.

Reviewed By: hansonw

Differential Revision: D3190244

fb-gh-sync-id: 4122eecbb2ecd0e59178b8d405bfc6d3d89d6393
fbshipit-source-id: 4122eecbb2ecd0e59178b8d405bfc6d3d89d6393