Clang server: Add an AST to JSON "get outline" command
Summary:This is a ghetto AST to JSON exporter we'll use to generate an outline view in Nuclide.

You can view the full list in the source code, but it currently supports:

- structs, unions, classes, enums
- Objective-C interfaces, categories, protcols, and implementations
- class templates & specializations
- namespaces
- functions, [de]constructors, methods, and Objective-C methods
  - parameters and template parameters are also recorded for these
- variables, fields, and Objective-C ivars
  - types are recorded for these

The only major thing that's missing is typedefs - mostly because I'm not sure what the best UI would be for them. In particular, it's unclear how to distinguish between typedefs and variables.

Output format is pretty straightforward:

```
{
  name: string,
  cursor_kind: string,
  cursor_type?: string,
  extent: {
    start: {line: int, column: int},
    end: {line: int, column: int},
  },
  params?: [string],
  tparams?: [string],
  children?: [this],
}
```

Reviewed By: matbd

Differential Revision: D3199623

fb-gh-sync-id: ea32d6aa5b188433c9a715afd7aa59d6a7dd83d1
fbshipit-source-id: ea32d6aa5b188433c9a715afd7aa59d6a7dd83d1