Refactor nuclide-node-transpiler for custom transform use
Summary:This is a complete re-write of `nuclide-node-transpiler`. It still provides the same functionality as before but now in more babel-module agnostic way. The main goal of this is to have a central place where the configuration for transpiling Nuclide code can live. And to share that configuration with the various independent systems that do transpiling.

The major change here is where the cache is stored for `transformWithCache` and the `require()` hook. Before it was stored in the package directory under `babel-cache`. Now it's stored in `$TMP/nuclide-node-transpiler`. In production environments, where the package may be read-only, it makes sense to put the cache in a temp dir.

  - `nuclide-task` => `nuclide-path-search`. `nuclide-task` spawns processes. `nuclide-path-search` is the only package that uses `nuclide-task`.
  - `nuclide-debugger-atom`. It runs in webview, but Atom doesn't provide those with transpiling.
  - `nuclide-jasmine`. To run node unit tests.
  - `nuclide-server`.

  - `nuclide-service-parser`, to transpile the generated marshalling proxies - although they run in Atom and on the server, the proxies aren't real files, so they can't take advantage of the require hook.

  - `lib/internal/atom-babel-compiler-patcher.js` to override Atom's built-in transforms with our own.
  - `scripts/dev/release_transpile.js` to transpile what gets published.

Reviewed By: hansonw

Differential Revision: D3206989

fb-gh-sync-id: 74d5b43ed962a3a6272830556f46593003878b7b
fbshipit-source-id: 74d5b43ed962a3a6272830556f46593003878b7b