Clang service: Add getOutline function
Summary:This adds a `getOutline` function to get the JSON outline data for a C++ source file.

It's okay if `getOutline` takes a long time, so I added a `blocking` flag to `makeRequest` to allow waiting for any existing compilation commands to finish. (Normally for autocomplete/hyperclick requests, we fail fast to allow fallbacks to methods like ctags.)

Reviewed By: peterhal

Differential Revision: D3199684

fb-gh-sync-id: e807fa654e6c5d00ed88e96a2f37b4ef399919d3
fbshipit-source-id: e807fa654e6c5d00ed88e96a2f37b4ef399919d3