Add a 'Shutdown Nuclide Servers' command
Summary:People frequently run into issues where the best solution is to restart both the Nuclide server and client. Telling people to "pkill -f nuclide" is a little unpleasant, so let's add a command for it.

Note that this preserves the workspace state by forcibly killing the servers without removing the connections (had to add some code for this.)

Finally, keep track of when / how frequently this is being used.

Reviewed By: zertosh

Differential Revision: D3210351

fb-gh-sync-id: 59525f87ca56c6f6294bfe935e42698eaa0487e6
fbshipit-source-id: 59525f87ca56c6f6294bfe935e42698eaa0487e6