Guard against multiple disposals
Summary:The motivation for this diff is that for a week or so I've been seeing a red logline in the console whenever we kill the php debugger.  This diff addresses this, as well as a few race conditions during cleanup of the debugger.

First, we have ConnectionMultiplexer's `dispose` method.  This diff changes the order so we wait on the script process promise to resolve BEFORE we remove all the connections.  The reason for doing this is that removing all the connections will always put the debugger into STATUS_END state.  Before this diff, we would then hit the `if` guard

`if (this._scriptLaunchPath != null && this._status === STATUS_END) {`

So if we happened to launch a script that took a while to run, and then decided to kill the debugger before it finished, we would still end up `await`ing the script launch process promise.  This is bad for a few reasons, but one of them is that the remote object's `dispose` method becomes arbitrarily slow, allowing it to be killed by the server when we reach the server's 60 second timeout, which would be a problem if our script would take over 60 seconds to run.

Second, we have guards against re-entering `dispose` methods.  `HhvmDebuggerProxyService`'s `dispose` method is called at least three times when you hit the red 'x' to kill the debugger.  Once from `DebuggerActions`, once from a callback made when the socket for the chrome UI ends (the callback is in `HhvmDebuggerInstance`) and once from within the server itself, once the state machine is put into the STATUS_END state.  Before this diff, we were suffering from a problem where the client-side would unsubscribe from its server messages observable, before the server-side had a chance to send all of it's messages.  Actually there was only one message, and it is the exit code message sent after a script finishes.  This was causing the red logline in the chrome devtools whenever you kill the debugger.  Since we are potentially calling `dispose` on the remote object from the client side multiple times, we need to guard against re-entry, otherwise we are at risk for putting the debugger into a bad state.

There is one last thing, which happens when you run a script under the debugger, and then kill the debugger before the script finishes.  In this case we will still try to send the exit code message, but it will most likely happen after the observable streams have been unsubscribed from.  Since calling `.next` on a stream with no subscribers throws an exception, I handle this with a try/catch.

Reviewed By: yinghuitan

Differential Revision: D3205330

fb-gh-sync-id: 0893cca10f9b5f33cf660f5d8ba3dabab65b636d
fbshipit-source-id: 0893cca10f9b5f33cf660f5d8ba3dabab65b636d