ActiveEditorBasedService
Summary:Adds an abstraction that allows people to easily create packages that fit the
following model:

- They render results computed from the contents of a text editor
- These results need to get recomputed whenever the currently-active text editor
  changes
- The actual thing that computes results depends on the grammar of the current
  file (e.g. different providers for Flow, Hack, etc.)

With this abstraction, such packages only need to subscribe to a result stream
and figure out how to render results.

In the future I would like to add options to allow clients to customize the
editor events they subscribe to. For example, some may want to recompute results
only on save rather than whenever the text is edited. We could even allow
clients to completely customize events by supplying a function from (TextEditor,
provider) to Observable, which would allow even something like the LinterAdapter
to be converted from text-event-dispatcher to this framework.

It may also be useful to add a facility to cache results for all of the open
tabs, rather than recomputing them on every tab switch. The
text-event-dispatcher was designed with that in mind, so for this to completely
subsume the text-event-dispatcher we will need similar functionality.

Reviewed By: peterhal

Differential Revision: D3179743

fb-gh-sync-id: 8c79f4989264eeca14cb0b403e4c748e7759f02f
fbshipit-source-id: 8c79f4989264eeca14cb0b403e4c748e7759f02f