Limit number of Flow workers
Summary:
This doesn't take hyperthreading into account (e.g. 4 cores presented as 8 due
to hyperthreading leads to 7 as the max-workers value), but Flow's heuristic
doesn't either so this is no worse.

The idea here is to leave at least 1 core available for non-flow things.
Anecdotally it seems to improve responsiveness somewhat. At the very least, it
doesn't break Flow support so it's worth trying.

Reviewed By: zertosh

Differential Revision: D3215618

fb-gh-sync-id: 2bf4863a634321d226b4e6f71c43a2baae1c20d0
fbshipit-source-id: 2bf4863a634321d226b4e6f71c43a2baae1c20d0