Selector in launch attach UI for project directory
Summary:
Adds a selector to the php debugger launch attach UI for selecting which remote project we want to
debug.  This is really only used to determine the script used by the REPL for loading context, but before this
change, this directory would be set to '/' which would never load the right context.  Now we make the choice
explicit.

Reviewed By: jgebhardt

Differential Revision: D3216285

fb-gh-sync-id: 26bb64763d96cfb6359cb59256ff8abfac508860
fbshipit-source-id: 26bb64763d96cfb6359cb59256ff8abfac508860