Separate out Flow outline tests
Summary:
This is mostly helpful because Jasmine doesn't print useful diffs between
objects when they are expected to be equal but aren't. Instead, it just dumps
the entirety of both objects and you have to stare at them to find the
differences, or use some other tool to compare them. Therefore it is helpful to
have several smaller test cases rather than one large one, especially as the
number of outline features grows. That way at least you have a pair of
relatively small objects to stare at.

Reviewed By: shushz

Differential Revision: D3221987

fb-gh-sync-id: eeb2e6a2d65b7b99c1d6348f0e298e0fa607e1b5
fbshipit-source-id: eeb2e6a2d65b7b99c1d6348f0e298e0fa607e1b5