Set up two-way communication over the UNIX socket to ack when data is written to an output stream.
Summary:
Previously, sometimes the output from `atom-script` would be truncated because
the Atom process would exit before it had finished writing back to `atom-script`.
This introduces a protocol where the Atom process sends a JSON blob to the main
process that contains:

* The data to be written.
* Which method of `console` is doing the writing.
* A unique message ID.

Once the main process has written the data to stdout or stderr, as appropriate,
it sends a message with the ID back to the Atom process to acknowledge that the
message has been written. The Atom process does not exit until it receives an ack
for all of its messages from the main process.

Reviewed By: zertosh

Differential Revision: D3204846

fb-gh-sync-id: 5361d0cfdb6316766974e424e8bacf7ba6ee01cd
fbshipit-source-id: 5361d0cfdb6316766974e424e8bacf7ba6ee01cd