Optimize Clang autocomplete: use a heap (2/n)
Summary:
Clang's autocompletion works by giving it a location after a punctuation character (e.g. after "a." or "a->") - it will then give you all results for that location, leaving filtering to the client.

My previous implementation works fine but it can be more efficient. Clang often returns 30-100k results for an autocompletion query, but we only display 100 of them!

The first optimization is to avoid sorting the entire result set - instead, we can filter and use a heap to select the top 100 results. This creates a new `CompletionResults` wrapper to make this code a bit cleaner.

Reviewed By: matbd

Differential Revision: D3214492

fb-gh-sync-id: b0e8d9e5b8ceb04d969324fe2dc8f3c90e1a87dd
fbshipit-source-id: b0e8d9e5b8ceb04d969324fe2dc8f3c90e1a87dd