Memoize `isRunningInTest` to prevent sneaky IPC
Summary:
Prior to Atom v1.7.0, `atom.inSpecMode` had a chance of performing an IPC call that could be
expensive depending on how much work the other process was doing. Because this value will not
change during run time, memoize the value to ensure the IPC call is performed only once.

See [`getWindowLoadSettings`][1] for the sneaky getter and `remote` call that this memoization
ensures happens only once.

[1]: https://github.com/atom/atom/blob/v1.6.2/src/window-load-settings-helpers.coffee#L10-L14

Reviewed By: zertosh

Differential Revision: D3244189

fb-gh-sync-id: 7385324bc9f5970445dd41586d27db1d0fc45ec1
fbshipit-source-id: 7385324bc9f5970445dd41586d27db1d0fc45ec1