Introduce and use passesGKSafe
Summary:
Currently, e.g. datatips and debugger toggling don't work offline or off VPN due to the GK check throwing. We should actually swallow the exception with a timeout of 0 - if the GK has loaded, great, otherwise it's fine if we return false.

Introduce a new commons function called `passesGKSafe` that defaults to false on exception (this should be used most of the time).

Reviewed By: jgebhardt

Differential Revision: D3243933

fb-gh-sync-id: 6b82d35f44b13006b8c0b4103450db5f9fa2bd05
fbshipit-source-id: 6b82d35f44b13006b8c0b4103450db5f9fa2bd05