Fix major issue with remote project reloading
Summary:
One long-standing quirk with remote projects is that Atom opens a new file with the path of the remote project directory on startup. While a bit annoying, this typically isn't a problem.

However, yinghuitan recently added the ability to open arbitrary files from the remote filesystem. This means that if the remote file URI handler fires before the remote connection can be re-established, Nuclide will attempt to open the remote directory as a file, permanently breaking our internal state and leaving the user unable to connect to that particular project!

If we open a file before the connection loads, we should always just open a blank editor and allow the `onDidAddRemoteConnection` handler take care of the rest.

Additionally, we can also solve the issue with the blank editors by creating a 'self-destructing editor'. It's a bit of a hack, but seems to work well.

Reviewed By: zertosh

Differential Revision: D3246684

fb-gh-sync-id: ac25294682d5317aff24dfc79a1b33a8d7a5bb3c
fbshipit-source-id: ac25294682d5317aff24dfc79a1b33a8d7a5bb3c