Fix incorrect conflicts with simultaneous remote file saves
Summary:
When network latency is high, it's possible to send save requests in a way that we'll incorrectly report conflicts on disk. Here's a timeline diagram to explain what happens:

```
|---save1: a---|
      |---save2: ab---|
          |---save3: abc---|
                 |---updateCachedDiskContents---|
                 x                              y
```

At time x, saves 2 and 3 haven't gone through so `cachedDiskContents` is still 'a'.
However, by the time `updateCachedDiskContents` finishes at time y, saves 2 and 3 do actually finish so the new `cachedDiskContents` is 'abc', which is correct. We incorrectly compare this against the outdated 'a' which triggers a conflict message.

We need to skip the conflict check if any saves completed while `updateCachedDiskContents` was executing, since these automatically invalidate `cachedDiskContents`. We can implement this via a simple incrementing `saveID` counter that goes up every time a save request finishes.

Presumably save2 and save3 will trigger more Watchman changes, so the conflict check will still happen - just after the last save.

(alternative: we could get `updateCachedDiskContents` to return the former contents when it overwrites them. that seems trickier though)

Side note: it's possible that an external change actually does come through while some save requests are going through, which we won't catch. However, if that's the case, we're already screwed since we can't cancel the save request - we lose the external change regardless. Thus I don't think this is an actual problem.

Reviewed By: zertosh

Differential Revision: D3246749

fb-gh-sync-id: b3ce05422e0ef597aa0ef7fb66505f61630d2328
fbshipit-source-id: b3ce05422e0ef597aa0ef7fb66505f61630d2328