Fix failed watches on new remote files/directories
Summary:
When creating a new file/folder, we sometimes fail to subscribe: the initial `ServerConnection.createFile` attempts to subscribe immediately, which of course fails since the file doesn't exist yet.

However, we never null out `watchSubscription`, so we never end up re-attempting to subscribe to the file when it's actually created in `create`!
This means that we never get updates to the file. In addition, the ServerConnection's entry never gets cleared so attempting to create a directory with the same name afterwards leads to an unclearable error!

This nulls out the observable subscription on failure or completion, allowing us to retry as necessary.

Reviewed By: mostafaeweda

Differential Revision: D3247134

fb-gh-sync-id: a21e80f3a0895d17638616fba3d42d374511fb40
fbshipit-source-id: a21e80f3a0895d17638616fba3d42d374511fb40