Diff View: Fix race condition between activating and revisions loading
Summary:
When the `DiffTimelineView` is created the second time, the revisions have already been loaded, but the `DiffViewModel` is still inactive.
Hence, calling `getActiveRevisionsState` in the constructor returns null.
For that, we need to activate the model before creating/rendering the UI components.
With that in mind, `DiffTimelineView` should get the state when it's mounted (not created), in this cases, done together.

Reviewed By: jamesgpearce

Differential Revision: D3251445

fb-gh-sync-id: 35d067e9ee3ab991e89ff6a28627b10b0580b3e9
fbshipit-source-id: 35d067e9ee3ab991e89ff6a28627b10b0580b3e9