nuclide-type-coverage package
Summary:
This package renders a status bar widget that displays how much of the current
file is covered by the type system. This is important information when working
with a gradually-typed type system such as Flow or Hack, since without it there
are cases where you may believe you have typechecking when really you do not.

In the future this package will also provide a way to see which particular parts
of your file are not covered. Obviously that is more actionable than an overall
percentage.

Reviewed By: peterhal

Differential Revision: D3250952

fb-gh-sync-id: c44c59f73671c0c3f386f414581af813661f1b9e
fbshipit-source-id: c44c59f73671c0c3f386f414581af813661f1b9e