Show 'nuclide-file-tree' in side-bar immediately if nothing else
Summary:
If a user previously opened/closed Atom on a project without firing an
event that affected the file tree's visibility, they will have `null`
serialized as their active view. In that case, ensure
'nuclide-file-tree' takes precedence since that's what everyone is
accustomed to.

This is a special case that is okay leaking between packages. The
file-tree is a special star.

Reviewed By: matthewwithanm

Differential Revision: D3244546

fb-gh-sync-id: 21c514563334b9ff270bb574318dbb2f2353c7a4
fbshipit-source-id: 21c514563334b9ff270bb574318dbb2f2353c7a4