Defend against possibly not being sent properties on evaluation
Summary: Last night I think I was running into an issue where no properties were being sent back as a response from an eval call to the debugger. I can't repro it now but Xdebug says this is a legal thing, so this code defends against this possibility.  I've added some logging to better investigate if it happens again, and we will just return null in this case.

Reviewed By: yinghuitan

Differential Revision: D3175163

fb-gh-sync-id: b1c0504fc00406fe5a3cda8d7dd18e642f74a8ea
fbshipit-source-id: b1c0504fc00406fe5a3cda8d7dd18e642f74a8ea