Add WatchExpressionListStore.
Summary: Stores the expressions listed in the watch expressions window. Unfortunately the better name `WatchExpressionStore` is already taken by the store managing watch expression requests in general (which are also used by e.g. debugger datatips). Open for suggestions, or I might rename this in the future if confusion arises.

Reviewed By: johnislarry

Differential Revision: D3242792

fb-gh-sync-id: 82a97a3d61fb6c892774fd33dc8f0ddddbbc110b
fbshipit-source-id: 82a97a3d61fb6c892774fd33dc8f0ddddbbc110b