Add new linter header type for flow only files
Summary:
This new header is meant for files that are read by flow but not by the transpiler. The build transpiler will transpile any file that starts with the right pragma (e.g. `'use babel'`, `"use babel"`, `/* flow */`, or `/** babel */`). These is some flow syntax that babel can't handle, this header is for those files.

Also changed the flow libs to use the new header, even though flow doesn't care about the `flow` pragma when it's a lib – but just to keep things consistent.

Reviewed By: matthewwithanm

Differential Revision: D3268722

fb-gh-sync-id: e46f6badc2bed22699cbd5339d00aa7c89c07863
fbshipit-source-id: e46f6badc2bed22699cbd5339d00aa7c89c07863