HHVM toolbar serialize script path
Summary: If you launch a script from the HHVM toolbar you can specify arguments by typing them into the text area.  If you switch tabs, any arguments you typed in will vanish, which is a poor experience if you want to run a script multiple times with lots of arguments.  This change allows the toolbar to remember the arguments when tabs are switched.

Reviewed By: yinghuitan

Differential Revision: D3251774

fb-gh-sync-id: 78372131eddf2502d79b55765e654dc3b5ef79f2
fbshipit-source-id: 78372131eddf2502d79b55765e654dc3b5ef79f2