Diff View: Fix fast editing/saving with long-delay saves reverting user changes and jumping cursor
Summary:
Previously, we only considered the latest saved contents to compare against the filesystem contents, which in long-delay save requests and fast editing/saving leads to reverting user changes and jumping the cursor to the end of the document.
Here, I'm introducing a self-managed cache of 20 saved contents for the active file to check against past saves.

Reviewed By: peterhal

Differential Revision: D3266088

fbshipit-source-id: cb8db49ddbef016776b94353bc8187bc6c73675c