Make InfoService.closeConnection asynchronous
Summary:
Since InfoService.closeConnection doesn't return a Promise, awaiting it doesn't actually wait for the socket message to go through - it seems to be more of a 'fire and forget'.

Make it asynchronous and mark the return type as `Promise<void>` so we can actually wait for the server to receive the message. This means we have to actually close the connection in a callback, as otherwise the response would never go through.

Reviewed By: zertosh

Differential Revision: D3279980

fbshipit-source-id: c5d6aa337d7f9ea969efa0054507463c2ba67f19