Add (disabled) no-re-export-type rule
Summary:
From the rules intro:

This rule is here because of a babel 5 bug, where "export type" is treated as a value export. This results in the types file getting required. In most cases, the types file gets transpiled to a mostly empty file (because types get removed), so there's only a small perf cost. However, there is some flow
syntax that is not supported by babel 5 (e.g. "export interface") that results in the transpiler throwing at runtime.

Reviewed By: hansonw

Differential Revision: D3273699

fbshipit-source-id: e5a9b5e48dbc4d4600847df8bb168a631ba54ef2