Debounce quick selection loading spinner and results list
Summary:
Reduce the DOM re-render flicker of quick selection search by:

1) Debouncing the results changed emit, such that provider search results are accumulated in short timespans rather than triggering multiple re-renders.
2) Only fire the loading event/spinner if the results list promise does not resolve before a time upper bound

The time bounds are currently constants -- when the bound is too small, flicker will still be visible (i.e. with 100ms debounce, there is no flicker with local search but there is with remote search). When the time bound is too long, there will be a noticeable delay before results appear.

Reviewed By: hansonw

Differential Revision: D3260838

fbshipit-source-id: 2c047f3dc91bbfa0388126b0fb5696bba5c9f628