Clang: attempt to find source files that include a header file
Summary:
fbcode developers often don't update their TARGETS files to include headers (or in some cases you have header-only targets).

To circumvent this, we can try searching for source files that include a header file and just use flags from there. The `findIncludingSourceFile` function generates a regular expression for includes and does a recursive grep in the header file's directory.

We handle the two most common types of include statements:

1. Includes relative to the project root (if supplied); e.g. `#include <a/b.h>`
2. Includes relative to the source file; e.g. `#include "../../a.h"`

In the second case, we're careful to check that the relative path actually leads to the header file in question.

We also use an Observable to enable cancellation after a timeout (15s), since directories could potentially be very large.

Reviewed By: matthewwithanm

Differential Revision: D3271769

fbshipit-source-id: cc71da51b24f0fae1d3b3fa3583631645e64e263