add integrateDelete
Summary:
When deleting we just toggle the visibility.  The tricky part is that we need to merge runs of deleted characters.  If you have the string "abc", and you delete a then c then b you should end up with a single run of deleted characters.  This means you have to look right and left to see if you need to merge runs.

There end up being 9 cases you have to worry about:
 (left edge and can merge left, left edge and can't merge left, not left edge) X (right edge can merge right, right edge can't merge right, not right edge)

As such there are 9 unit tests for these cases.

Reviewed By: nmote

Differential Revision: D3043283

fbshipit-source-id: 3a10bbcc1d02b6d133e8d2851b1a826cd9bd01bc