add pos method
Summary: This adds the ability to get the position of a WChar with is an important operation for later methods.  This method is complicated by the fact that you can either ignore invisible characters, or include them.  The second paramter to the function specified whether you want to ignore invisibles when calculating the offset.

Reviewed By: nmote

Differential Revision: D3043423

fbshipit-source-id: 05fc95bcba074919baa3585d14e65d40162bdf4a