add visible ranges method
Summary:
visibleRanges takes a list of runs and returns the list of visible ranges in those runs.  This is important for generating a list of changes that are the result of a delete operation.

For example consider 2 equal WStrings:  1: abc, 2: abc.  Say 1 deletes b.  And 2 deletes the whole string.  This could happen:

1: abc => ab*c
2: abc => a*b*c* creates op2
1: integrates op2.  To determine which characters to remove we must find the ranges for a and b are visible and remove them.

Reviewed By: nmote

Differential Revision: D3044110

fbshipit-source-id: aea3f947c07572523970c3d2190f5acf2fb75760