Quick Open: tone down debounce and loading delays
Summary:
Once we can display cached results while waiting, the loading delay can be reduced to 200ms and the re-render debounce to 50ms.

Also revert a change that defaults query debounce to 200ms in the component (this is handled in SearchResultManager already).
I also found this unused `_hasNoResults` function that looks like it can be safely removed.

Reviewed By: tyangliu

Differential Revision: D3288581

fbshipit-source-id: 336445b820103d4ea75fb034d1208e93ba2a1567