Quick Open: show previously cached results while waiting for new ones
Summary:
Now that we wait to show the loading indicators (thanks to tyangliu), there are brief windows of time in which omnisearch yields no results (because nothing has come back yet, not even a loading indicator).

Instead of showing nothing, let's keep track of the last cached result and display that if we have nothing else.

Reviewed By: ssorallen

Differential Revision: D3288530

fbshipit-source-id: 3e346740b411e2bc51e3b6367e46df2c074bb71a