Commons: checkOutput should not throw on ENOENT
Summary:
The wrongly-named `checkOutput` currently rejects with an Object on errors that prevent the process from starting such as ENOENT. A thorough examination of all the callsites leads me to believe that nobody wants this; very few callers are actually checking for exceptions.

Instead, just resolve with a null exitCode. Most callers do actually check if the exitCode is 0, so this doesn't break anything.

Reviewed By: zertosh

Differential Revision: D3288998

fbshipit-source-id: 5d9a9d23e59539d385c018724d969f1e5ddc6e26