Commons: Make asyncExecute throw a real Error
Summary:
`asyncExecute` currently throws an Object with the exit code, stdout, and stderr if the exit code is non-zero. This is madness, and makes test failures impossible to debug (they always say "Object object")

Instead, throw a normal error. Any callers who needed the error's properties should use `checkOutput` and check the exit code themselves. (The naming here is also wrong, but that's for another time...)

Reviewed By: zertosh

Differential Revision: D3286287

fbshipit-source-id: a4028c1a5d37a07d5f3eaa4d416ea86e8383daaf