Suppress chromium noise from test output
Summary:
This diff suppresses the lines below from the streaming test output, but keeps them in the stdout reprint when there's a test failure. I left "503" and "305" in the regex so when we update Atom, and those line numbers change, we're forced to re-evaluate them.

```
[52408:0511/165303:WARNING:resource_bundle.cc(503)] locale resources are not loaded
[52404:0511/165234:WARNING:resource_bundle.cc(305)] locale_file_path.empty() for locale English
```

I couldn't find any definitive answer on what these errors are. If you Google them, they show up repeatedly in Chromium based projects. Nobody is complaining about it, it's usually just part of some other output. I'm guess that it's "normal" since results go back to 2014.

Reviewed By: hansonw

Differential Revision: D3291405

fbshipit-source-id: 1faa09b32d6827f21cabf8e49581c6a8179e55d0