Add custom test runner for integration tests
Summary:
This is the start of a custom test runner for integration tests. Right now it's mostly just a passthrough to the built-in "legacy test runner". Noteworthy points:

* `configDirPath` is now set by the test runner, not in the UP loader.
* 3rd-party packages are never installed by Atom during integration tests. Instead, `js_test_runner.py` installs `tool-bar` before running the tests.
* The test runner handles temp dir cleanup, instead of leaving to the process exit event. Atom overrides this, so temp dirs never actually got cleaned up.

For the most part, most of the `nuclide-integration-test-helpers` startup boilerplate could be moved into the test runner, except `jasmine.useRealClock()`. Since we still delegate to the legacy test runner, it'll undo that.

The next step is to get the bits we need from https://github.com/atom/atom/blob/master/spec/spec-helper.coffee so we can stop delegating to the legacy test runner. Then, most of the setup done by `nuclide-integration-test-helpers` will be automatic for integration tests.

Reviewed By: johnislarry

Differential Revision: D3283620

fbshipit-source-id: 06de85e8979579d9c27028c49df27f196879db2e