Default all <AtomInput /> components to be in tab order
Summary:
The `<atom-text-editor>` created by `AtomInput` should be in tab
order by default for keyboard accessibility. Giving it a tabindex of
"0" puts it in tab order. "0" behaves the same as "-1" in terms of
programmatic focusability but also inserts the element into the tab
order of the document, making it accessible by the user pressing 'tab'.

Known issue: 'shift-tab' does not cycle backwards out of a focused
atom-text-editor. This is an issue in the custom component that must be
addressed in core Atom.

Reviewed By: jgebhardt

Differential Revision: D3287996

fbshipit-source-id: f87743d32916e4f5a709ae3afd49f2e34825f218